Implementation and Specifications for ProjectionKey to EventSourceId converter.
We can use a 128 bit hashing function since it is a one way conversion
from string to Guid. Since it is not used for hiding/verifying secret
information, it should be a non-cryptographic hash. The CityHash was
chosen because it is optimised for string hashing - and according to
the documentation should be independent of platform endianness.