Remove IWaitForAggregateRootEvents.
It is an optimisation that we don't need right now. We can just
check every time an event is committed. It's going to cost a bit of CPU
but it turns out to be a lot of work to implement with the current
abstractions. We'll get back to this later.