Change pings to be written "synchronously"
Ideally we want all of it to happen on the same thread, but as the gRPC
library exposes an WriteAsync method, we cannot guarantee that all the
execution happens on the same thread. This at leas performs as much as
possible on the calling thread. The current implementation in the
gRPC library also does most of the work synchronously, only delegating
the actual write to the server running in a different thread (which we
can't change). So this should work nicely for us.